Refatoração de arquivo do resolvedor.
